Moved fetchApiResponseData into helpers subfolder.
